rename format package to outpututils
